@@ -1,14 +1,27 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
 import { LoginComponent } from './login.component';
+import { Routes } from '@angular/router';
+import { RouterTestingModule } from '@angular/router/testing';
+import { HttpClientModule } from '@angular/common/http';
+import { AuthService } from '../../services/auth/auth.service';
 
 describe('LoginComponent', () => {
   let component: LoginComponent;
   let fixture: ComponentFixture<LoginComponent>;
+  const routes: Routes = [
+    {path: 'login', component: LoginComponent}
+  ];
+  let authService: AuthService;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ LoginComponent ]
+      declarations: [
+        LoginComponent
+      ],
+      imports: [
+        RouterTestingModule.withRoutes(routes),
+        HttpClientModule
+      ]
     })
     .compileComponents();
   }));
@@ -16,10 +29,20 @@ describe('LoginComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(LoginComponent);
     component = fixture.componentInstance;
-    fixture.detectChanges();
+    authService = TestBed.get(AuthService);
   });
 
-  it('should create', () => {
+  afterEach(() => {
+    authService = null;
+  });
+
+  it('should create login component', () => {
     expect(component).toBeTruthy();
+  });
+
+  it('should check the login method', () => {
+    const spy = spyOn(authService, 'login');
+    component.login();
+    expect(spy).toHaveBeenCalled();
   });
 });
