@@ -1,11 +1,17 @@
-import { Component } from '@angular/core';
-import {switchMap} from 'rxjs/internal/operators';
-import {of, concat} from 'rxjs';
-import {SpotifyService} from '../../services/spotify/spotify.service';
-import {PlaylistService} from '../../services/playlist/playlist.service';
-import {StatusBarService} from '../../services/status-bar/status-bar.service';
-import {animate, state, style, transition, trigger} from '@angular/animations';
-import {TrackService} from '../../services/track/track.service';
+import { Component, OnInit } from '@angular/core';
+import { switchMap } from 'rxjs/operators';
+import { concat, of } from 'rxjs';
+import { SpotifyService } from '../../services/spotify/spotify.service';
+import { PlaylistService } from '../../services/playlist/playlist.service';
+import { StatusBarService } from '../../services/status-bar/status-bar.service';
+import { animate, state, style, transition, trigger } from '@angular/animations';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
+import { NewPlaylistDialogComponent } from '../new-playlist-dialog/new-playlist-dialog.component';
+import { Router } from '@angular/router';
+import { SpotifyPlaylistRespose } from '../../interfaces/playlist/spotifyPlaylistResponse.interface';
+import { CurrentTrack } from '../../interfaces/track/current-track.interface';
+import { UtilService } from '../../services/util/util.service';
+import { Playlist } from '../../interfaces/playlist/playlist.interface';
 
 @Component({
   selector: 'app-spotify-navigation-menu',
@@ -24,96 +30,93 @@ import {TrackService} from '../../services/track/track.service';
     ])
   ]
 })
-export class SpotifyNavigationMenuComponent {
-  public playlists: Array<Object> = [];
+export class SpotifyNavigationMenuComponent implements OnInit {
+  public playlists: Array<Playlist> = [];
   public loading: boolean;
   public playlistsLoaded: boolean;
-  private selectedPlaylist: string;
+  public selectedPlaylist: string;
   public isPictureEnlarged: boolean;
-  public currentTrack: Object;
+  public currentTrack: CurrentTrack;
   public imageEnlargeState: string;
 
-  constructor(private spotifyService: SpotifyService, private playlistService: PlaylistService, private statusBarService: StatusBarService, private trackService: TrackService) {
-    let token = '';
+  constructor(
+    private spotifyService: SpotifyService,
+    private playlistService: PlaylistService,
+    private statusBarService: StatusBarService,
+    public dialog: MatDialog,
+    private router: Router,
+    public utilService: UtilService) {}
+
+  ngOnInit() {
     this.selectedPlaylist = '';
     this.imageEnlargeState = 'inactive';
     this.isPictureEnlarged = false;
-    this.statusBarService.enlargePicture$.subscribe(value => {
+    this.statusBarService.enlargePicture$.subscribe((value: boolean) => {
       this.isPictureEnlarged = value;
       this.imageEnlargeState = value ? 'active' : 'inactive';
     });
-    this.statusBarService.currenttrack$.subscribe(value => this.currentTrack = value['track']);
-    this.trackService.currentTrack$.subscribe(value => this.currentTrack = value['track_window']['current_track']);
-    this.spotifyService.getAuthToken()
+    this.statusBarService.currentTrack$.subscribe((value: CurrentTrack) => this.currentTrack = value);
+    this.playlistService.selectPlaylist$.subscribe((playlist: string) => this.selectedPlaylist = playlist);
+    this.spotifyService.getAllPlaylists()
       .pipe(
-        switchMap(spotifyToken => {
-          token = spotifyToken['token'];
-          const isLoggedIn = !!token;
-          if (isLoggedIn) {
-            return this.spotifyService.getAllPlaylists(token);
-          } else {
-            this.playlists = [];
-            this.playlistsLoaded = false;
-            return of();
-          }
-        }),
-        switchMap(playlistInfo => {
+        switchMap((playlistInfo: SpotifyPlaylistRespose) => {
           this.loading = true;
           this.playlistsLoaded = false;
-          const isLoggedIn = !!token;
-          const tempList = [];
-          if (isLoggedIn) {
-            const playlistsLength = playlistInfo['total'];
-            const owner = String(playlistInfo['next']).split('users/')[1].split('/playlists')[0];
+          if (playlistInfo.items.length > 0) {
+            const tempList = [];
+            const playlistsLength = playlistInfo.total;
+            const owner = String(playlistInfo.next).split('users/')[1].split('/playlists')[0];
             const numberOfTimesToLoop = Math.ceil(playlistsLength / 50);
             for (let i = 0; i < numberOfTimesToLoop; i++) {
               const baseURI = `https://api.spotify.com/v1/users/${owner}/playlists?offset=${i * 50}&limit=50`;
-              tempList.push(this.spotifyService.getAllPlaylists(token, baseURI));
+              tempList.push(this.spotifyService.getAllPlaylists(baseURI));
             }
             return concat(...tempList);
           } else {
             this.loading = false;
-            this.playlists = [];
+            this.playlistsLoaded = true;
             return of();
           }
         })
       )
-      .subscribe(data => {
-        data['items'].forEach(playlist => {
-          if (playlist['name'] === this.selectedPlaylist) {
-            playlist['selected'] = true;
+      .subscribe((data: SpotifyPlaylistRespose) => {
+        data.items.forEach((playlist: Playlist) => {
+          if (playlist.name === this.selectedPlaylist) {
+            playlist.selected = true;
           } else {
-            playlist['selected'] = false;
+            playlist.selected = false;
           }
         });
-        if (!data['next']) {
+        if (data.testing || !data.next) {
           this.loading = false;
           this.playlistsLoaded = true;
         }
-        this.playlists = this.playlists.concat(data['items']);
+        this.playlists = this.playlists.concat(data.items);
       });
-
-    this.playlistService.selectPlaylist$.subscribe(playlist => {
-      this.selectedPlaylist = playlist;
-    });
   }
 
-  goToTracks(playlist) {
+  goToTracks(playlist): void {
     this.playlists.forEach(tt => tt['selected'] = false);
     playlist['selected'] = true;
-    this.playlistService.savePlaylist(playlist);
+    const playlistName = encodeURI(playlist.name.toLowerCase());
+    const playlistId = encodeURI(playlist.id);
+    const url = this.utilService.encodeSpecialSymbols(`/playlist/${playlistName}/${playlistId}`);
+    this.router.navigateByUrl(url);
   }
 
-  shortenString(string) {
-    const stringLength = 25;
-    if (string.length > stringLength) {
-      return string.substr(0, stringLength) + '...';
-    } else {
-      return string;
-    }
+  goToSavedAlbums(): void {
+    this.router.navigate(['library/albums']);
   }
 
-  shrinkPicture() {
+  shrinkPicture(): void {
     this.statusBarService.enlargePicture(false);
+  }
+
+  openNewPlaylistModal(): void {
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.panelClass = 'new-playlist-panel';
+    dialogConfig.height = '480px';
+    dialogConfig.width = '800px';
+    this.dialog.open(NewPlaylistDialogComponent, dialogConfig);
   }
 }
