@@ -1,15 +1,52 @@
-import { TestBed, async, inject } from '@angular/core/testing';
-
+import { TestBed } from '@angular/core/testing';
 import { SpotifyGuard } from './spotify.guard';
+import { Routes, Router, RouterStateSnapshot, ActivatedRouteSnapshot } from '@angular/router';
+import { LoginComponent } from '../../components/login/login.component';
+import { RouterTestingModule } from '@angular/router/testing';
+import { HttpClientModule } from '@angular/common/http';
+import { AuthService } from '../../services/auth/auth.service';
 
 describe('SpotifyGuard', () => {
+  let spotifyGuard: SpotifyGuard;
+  let router: Router;
+  let authService: AuthService;
+  const routes: Routes = [
+    {path: 'login', component: LoginComponent},
+  ];
+  const mockSnapshot: any = jasmine.createSpyObj<RouterStateSnapshot>('RouterStateSnapshot', ['toString']);
+
   beforeEach(() => {
     TestBed.configureTestingModule({
-      providers: [SpotifyGuard]
+      imports: [
+        HttpClientModule,
+        RouterTestingModule.withRoutes(routes)
+      ],
+      providers: [
+        SpotifyGuard,
+        {provide: RouterStateSnapshot, useValue: mockSnapshot}
+      ]
     });
+
+    router = TestBed.get(Router);
+    authService = TestBed.get(AuthService);
   });
 
-  it('should ...', inject([SpotifyGuard], (guard: SpotifyGuard) => {
-    expect(guard).toBeTruthy();
-  }));
+  afterEach(() => {
+    router = null;
+    authService = null;
+  });
+
+  it('should create spotify guard with token', () => {
+    spyOn(authService, 'getSpotifyToken').and.returnValue('testCookie');
+    spotifyGuard = new SpotifyGuard(router, authService);
+    expect(spotifyGuard.canActivate(new ActivatedRouteSnapshot(), mockSnapshot)).toBeTruthy();
+  });
+
+  it('should create spotify guard without token', () => {
+    const spy = spyOn(router, 'navigate');
+    spyOn(authService, 'getSpotifyToken').and.returnValue('');
+    spotifyGuard = new SpotifyGuard(router, authService);
+    expect(spotifyGuard.canActivate(new ActivatedRouteSnapshot(), mockSnapshot)).toBeFalsy();
+    expect(spy).toHaveBeenCalled();
+  });
 });
