@@ -1,22 +1,21 @@
 import { Injectable } from '@angular/core';
-import {CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router} from '@angular/router';
+import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router } from '@angular/router';
 import { Observable } from 'rxjs';
-import {SpotifyService} from '../../services/spotify/spotify.service';
+import { AuthService } from '../../services/auth/auth.service';
 
 @Injectable({
   providedIn: 'root'
 })
 export class SpotifyGuard implements CanActivate {
-  constructor(private router: Router, private spotifyService: SpotifyService) {}
+  constructor(private router: Router, private authService: AuthService) {}
   canActivate(
     next: ActivatedRouteSnapshot,
     state: RouterStateSnapshot): Observable<boolean> | Promise<boolean> | boolean {
-    this.spotifyService.getAuthToken().subscribe(token => {
-      if (!token['token']) {
-        this.router.navigate(['']);
+      const token = !!this.authService.getSpotifyToken();
+      if (!token) {
+        this.router.navigate(['login']);
         return false;
       }
-    });
-    return true;
+      return true;
   }
 }
