@@ -1,4 +1,5 @@
 import { Pipe, PipeTransform } from '@angular/core';
+import { Track } from '../../interfaces/track/track.interface';
 
 @Pipe({
   name: 'filterTrackArtist'
@@ -20,17 +21,15 @@ export class FilterTrackArtistPipe implements PipeTransform {
       return artistString;
     });
   }
-  transform(tracks: any, args?: any): any {
+  transform(tracks: Array<Track>, args?: string): any {
     const that = this;
-    if (tracks) {
-      if (args) {
-        return tracks.filter(track => {
-          const artist = that.displayArtists(track.track.artists).join('');
-          return artist.toLowerCase().includes(args.toLowerCase());
-        });
-      } else {
-        return tracks;
-      }
+    if (args) {
+      return tracks.filter((track: Track) => {
+        const artist = that.displayArtists(track.artists).join('');
+        return artist.toLowerCase().includes(args.toLowerCase());
+      });
+    } else {
+      return tracks;
     }
   }
 }
