@@ -1,8 +1,213 @@
 import { FilterTrackNamePipe } from './filter-track-name.pipe';
+import { Track } from '../../interfaces/track/track.interface';
 
 describe('FilterTrackNamePipe', () => {
-  it('create an instance', () => {
+  it('create an instance filter track name pipe', () => {
     const pipe = new FilterTrackNamePipe();
     expect(pipe).toBeTruthy();
+  });
+
+  it('check transform method with args', () => {
+    const tracks: Array<Track> = [
+      {
+        album: {
+          album_type: '',
+          artists: [],
+          available_markets: [],
+          external_urls: {
+            spotify: ''
+          },
+          href: '',
+          id: '',
+          images: [],
+          name: '',
+          release_date: '',
+          release_date_precision: '',
+          total_track: 0,
+          type: '',
+          uri: ''
+        },
+        artists: [],
+        available_markets: [],
+        disc_number: 0,
+        duration_ms: 0,
+        explicit: true,
+        external_ids: {
+          isrc: ''
+        },
+        external_urls: {
+          spotify: ''
+        },
+        href: '',
+        id: '',
+        name: 'test1',
+        popularity: 0,
+        preview_url: '',
+        track_number: 0,
+        type: '',
+        uri: '',
+        isPlayButtonShowing: true,
+        isPauseButtonShowing: true,
+        remove: true,
+        album_name: '',
+        title: '',
+        artist: '',
+        time: '',
+        addedAt: '',
+        duration: 0
+      },
+      {
+        album: {
+          album_type: '',
+          artists: [],
+          available_markets: [],
+          external_urls: {
+            spotify: ''
+          },
+          href: '',
+          id: '',
+          images: [],
+          name: '',
+          release_date: '',
+          release_date_precision: '',
+          total_track: 0,
+          type: '',
+          uri: ''
+        },
+        artists: [],
+        available_markets: [],
+        disc_number: 0,
+        duration_ms: 0,
+        explicit: true,
+        external_ids: {
+          isrc: ''
+        },
+        external_urls: {
+          spotify: ''
+        },
+        href: '',
+        id: '',
+        name: 'lala',
+        popularity: 0,
+        preview_url: '',
+        track_number: 0,
+        type: '',
+        uri: '',
+        isPlayButtonShowing: true,
+        isPauseButtonShowing: true,
+        remove: true,
+        album_name: '',
+        title: '',
+        artist: '',
+        time: '',
+        addedAt: '',
+        duration: 0
+      }
+    ];
+    const pipe = new FilterTrackNamePipe();
+    const newTracks = pipe.transform(tracks, 'tes');
+    expect(newTracks.length).toEqual(1);
+  });
+
+  it('check transform method without args', () => {
+    const tracks: Array<Track> = [
+      {
+        album: {
+          album_type: '',
+          artists: [],
+          available_markets: [],
+          external_urls: {
+            spotify: ''
+          },
+          href: '',
+          id: '',
+          images: [],
+          name: '',
+          release_date: '',
+          release_date_precision: '',
+          total_track: 0,
+          type: '',
+          uri: ''
+        },
+        artists: [],
+        available_markets: [],
+        disc_number: 0,
+        duration_ms: 0,
+        explicit: true,
+        external_ids: {
+          isrc: ''
+        },
+        external_urls: {
+          spotify: ''
+        },
+        href: '',
+        id: '',
+        name: 'test1',
+        popularity: 0,
+        preview_url: '',
+        track_number: 0,
+        type: '',
+        uri: '',
+        isPlayButtonShowing: true,
+        isPauseButtonShowing: true,
+        remove: true,
+        album_name: '',
+        title: '',
+        artist: '',
+        time: '',
+        addedAt: '',
+        duration: 0
+      },
+      {
+        album: {
+          album_type: '',
+          artists: [],
+          available_markets: [],
+          external_urls: {
+            spotify: ''
+          },
+          href: '',
+          id: '',
+          images: [],
+          name: '',
+          release_date: '',
+          release_date_precision: '',
+          total_track: 0,
+          type: '',
+          uri: ''
+        },
+        artists: [],
+        available_markets: [],
+        disc_number: 0,
+        duration_ms: 0,
+        explicit: true,
+        external_ids: {
+          isrc: ''
+        },
+        external_urls: {
+          spotify: ''
+        },
+        href: '',
+        id: '',
+        name: 'lala',
+        popularity: 0,
+        preview_url: '',
+        track_number: 0,
+        type: '',
+        uri: '',
+        isPlayButtonShowing: true,
+        isPauseButtonShowing: true,
+        remove: true,
+        album_name: '',
+        title: '',
+        artist: '',
+        time: '',
+        addedAt: '',
+        duration: 0
+      }
+    ];
+    const pipe = new FilterTrackNamePipe();
+    const newTracks = pipe.transform(tracks);
+    expect(newTracks.length).toEqual(2);
   });
 });
